int cast added to VtkActorViewer::getSize()

+++ include/VtkActorViewer.h
@@ -63,7 +63,7 @@ public:
     void setSize( size_t width, size_t height);
     inline size_t getWidth() const { return _ren->GetSize()[0];}
     inline size_t getHeight() const { return _ren->GetSize()[1];}
-    inline cv::Size getSize() const { return cv::Size( getWidth(), getHeight());}
+    inline cv::Size getSize() const { return cv::Size( (int)getWidth(), (int)getHeight());}
 
     cv::Mat_<float> getRawZBuffer() const; // Grab snapshot of the raw Z buffer from the viewer.
     cv::Mat_<cv::Vec3b> getColourImg() const; // Grab snapshot of whatever's currently displayed (3 byte BGR order)
